Modify Push and add code for notifications
Modify method `sendMany` in Push so that it now sends only to an
iterable list.

Add code to send the push notifications for the various events to the
users who are assigned to the task. This is safe since the Push
Notifications are sent within a try catch block and the exceptions are
logged.

Add docs for the various push events.